Add Windows Battery Class. (OOP)
Windows battery class simplifies the Windows API used to get the battery
information. Using this you can get the battery percent, battery flag,
and is the pc plugged in. There are methods for providing many data
types, so in that case, people who use this wouldn't have to use
stringstream or other ways of conversion from one data type to another.